Remove old tokenizer
