--- frontend/src/metabase-lib/v1/expressions/tokenize/index.ts
@@ -1,1 +1,0 @@
-export { tokenize } from "./tokenize";
